Create new command line argument to define the module name.
On the previous version the module name was fixed as 'All' not allowing
the developer to specify what will be the module name will be used for
store the deployment commit hash.

It was added a new option -o to define what will be the module name. The
module name will be 'All' if no argument is specified.